Reverting to 463a51e9

+++ components/include/spi2.h
@@ -146,7 +146,7 @@ esp_err_t setRADDR(uint8_t * adr);
 
 bool data_ready();
 
-void Get_Data(uint8_t * reci_data, uint8_t payload);
+void Get_Data(int * reci_data, uint8_t payload);
 
 uint8_t GetFIFOStatus();
